+++ packages/client-ledger/RELEASES.md
@@ -17,3 +17,13 @@ Based on:
 - [typescript v1.0.0-alpha.1] ./packages/client-payments
 ### Releases
 - [NPM v1.0.0-alpha.1] https://www.npmjs.com/package/@formance/sdk-payments/v/1.0.0-alpha.1 - ./packages/client-payments
+
+## 2024-07-05 00:15:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.326.3 (2.359.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.2] packages/client-ledger
+### Releases
+- [NPM v1.0.0-alpha.2] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.2 - packages/client-ledger